Changed Query::ranges() to use double precision.
Note that GPU calculations will be made using single precision.

+++ ohm/Query.cpp
@@ -65,7 +65,7 @@ const Key *Query::intersectedVoxels() const
 }
 
 
-const float *Query::ranges() const
+const double *Query::ranges() const
 {
   return (!imp_->ranges.empty()) ? imp_->ranges.data() : nullptr;
 }
