+++ ohm/private/QueryDetail.h
@@ -24,8 +24,8 @@ struct ohm_API QueryDetail
   OccupancyMap *map = nullptr;
   /// The voxels intersected by the query, identified by their @c Key (if used).
   std::vector<Key> intersected_voxels;
-  /// Distances associated with the @c intersected_voxels (if used).
-  std::vector<float> ranges;
+  /// Distances associated with the query results (if used).
+  std::vector<double> ranges;
   /// Number of results in @c intersected_voxels and/or @c ranges .
   size_t number_of_results = 0;
   /// @c QueryFlag values for the query.
@@ -38,8 +38,8 @@ struct ohm_API QueryDetail
 /// Query result helper identifying the closest result index.
 struct ohm_API ClosestResult
 {
-  size_t index = 0;                                 ///< Index into the results arrays.
-  float range = std::numeric_limits<float>::max();  ///< (Closest) range value.
+  size_t index = 0;                                   ///< Index into the results arrays.
+  double range = std::numeric_limits<double>::max();  ///< (Closest) range value.
 };
 }  // namespace ohm
